backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-dotenv
openai
tiktoken
httpx
aiohttp
Pillow
python-multipart
python-docx
PyPDF2
langdetect